fix(tests): extract messageDisplay utils to avoid Ink WASM issues in CI
Move getMessageDisplayContent and getMessageTextColor functions to
src/utils/messageDisplay.ts to avoid importing Ink components during
tests. This fixes the yoga-layout WebAssembly error that occurs on
Linux CI environments.
