+++ src/components/ChatInterface.tsx
@@ -14,55 +14,8 @@ import ModelSelector from './ModelSelector';
 import NotesSelector from './NotesSelector';
 import type { QueuedMessage } from '../hooks/useChat';
 import { RequestLogDisplay } from './RequestLogDisplay';
-import { renderMarkdown, isASCIICapableTerminal } from '../utils/markdownRenderer';
 import { useScreenSize } from "fullscreen-ink";
-
-/**
- * Get the display content for a message.
- * Extracted as a pure function for testability.
- */
-export function getMessageDisplayContent(message: Message): string {
-  const isSystemMessage = message.role === 'system';
-  const isUINotification = message.role === 'ui-notification';
-  const isUserMessage = message.role === 'user';
-  const isAssistantMessage = message.role === 'assistant';
-
-  // System messages: add "System:" prefix, no markdown
-  if (isSystemMessage) {
-    return `System: ${message.content}`;
-  }
-
-  // User messages: add "> " prefix, no markdown
-  if (isUserMessage) {
-    return `> ${message.content}`;
-  }
-
-  // For assistant AND ui-notification messages, render markdown if supported
-  if ((isAssistantMessage || isUINotification) && isASCIICapableTerminal()) {
-    try {
-      return renderMarkdown(message.content);
-    } catch (error) {
-      console.error('Failed to render markdown:', error);
-      return message.content;
-    }
-  }
-
-  return message.content;
-}
-
-/**
- * Get the text color for a message based on its role.
- * Extracted as a pure function for testability.
- */
-export function getMessageTextColor(message: Message): string {
-  if (message.role === 'system' || message.role === 'ui-notification') {
-    return '#cb9774';
-  }
-  if (message.role === 'user') {
-    return 'white';
-  }
-  return 'brightWhite';
-}
+import { getMessageDisplayContent, getMessageTextColor } from '../utils/messageDisplay';
 
 interface ChatInterfaceProps {
   completedMessages: Message[];
